feat: extract PG support into umzeption-pg package
Sprint 27-30: complete PostgreSQL extraction:

Core package (umzeption):
- remove PG exports (createUmzeptionPgContext, UmzeptionPgStorage, pgInstallSchemaFromString)
- remove pg types from advanced-types.d.ts (DefineUmzeptionContexts now only has 'unknown')
- installSchemaFromString now uses a plugin registry (registerSchemaInstaller)
- DefineUmzeptionContexts is now extensible via module augmentation

umzeption-pg package:
- lib/context.js, storage.js, utils.js, main.js (moved with history)
- lib/schema-helper.js: pgInstallSchemaFromString + auto-registers with core
- lib/types.d.ts: FastifyPostgresStyleDb, UmzeptionPgContext, module augmentation
- index.js + index.d.ts: exports createUmzeptionPgContext, UmzeptionPgStorage, pgInstallSchemaFromString
- test/: pg-integration, pg-utils, schema-helper, storage tests (moved with history)

Update .gitignore to allow packages/*/index.d.ts and umzeption-pg/lib/types.d.ts.
