+++ .gitignore
@@ -14,5 +14,8 @@
 *.d.ts.map
 !/lib/**/*-types.d.ts
 !/index.d.ts
+!/packages/*/index.d.ts
+!/packages/*/lib/**/*-types.d.ts
+!/packages/umzeption-pg/lib/types.d.ts
 
 # Library specific ones
