+++ packages/umzeption-pg/index.d.ts
@@ -1,0 +1,16 @@
+export type {
+  FastifyPostgresStyleDb,
+  UmzeptionPgContext,
+} from './lib/types.d.ts';
+
+export {
+  createUmzeptionPgContext,
+} from './lib/context.js';
+
+export {
+  UmzeptionPgStorage,
+} from './lib/storage.js';
+
+export {
+  pgInstallSchemaFromString,
+} from './lib/schema-helper.js';
