+++ packages/umzeption-pg/lib/context.js
@@ -1,0 +1,14 @@
+import { createUmzeptionContext } from 'umzeption';
+import { createFastifyPostgresStyleDb } from './utils.js';
+
+/** @typedef {import('./types.d.ts').FastifyPostgresStyleDb} FastifyPostgresStyleDb */
+
+/**
+ * @param {FastifyPostgresStyleDb["pool"]} pool
+ * @returns {import('umzeption').UmzeptionContext<'pg', FastifyPostgresStyleDb>}
+ */
+export function createUmzeptionPgContext (pool) {
+  const db = createFastifyPostgresStyleDb(pool);
+
+  return createUmzeptionContext('pg', db);
+}
