+++ packages/umzeption-pg/test/pg-integration.spec.js
@@ -6,11 +6,12 @@ import sinon from 'sinon';
 import pg from 'pg';
 import { Umzug } from 'umzug';
 
-import { UmzeptionPgStorage, createUmzeptionPgContext, umzeption } from '../index.js';
+import { umzeption } from 'umzeption';
+import { UmzeptionPgStorage, createUmzeptionPgContext } from '../index.js';
 
-import { down as downMain, up as upMain } from './fixtures/migrations/foo-01.js';
-import { installSchema as installSchemaTestDependency } from './fixtures/test-dependency/index.js';
-import { down as downTestDependency, up as upTestDependency } from './fixtures/test-dependency/migrations/foo-01.js';
+import { down as downMain, up as upMain } from '../../umzeption/test/fixtures/migrations/foo-01.js';
+import { installSchema as installSchemaTestDependency } from '../../umzeption/test/fixtures/test-dependency/index.js';
+import { down as downTestDependency, up as upTestDependency } from '../../umzeption/test/fixtures/test-dependency/migrations/foo-01.js';
 
 function getDependencyStubCallCount ({
   down = 0,
@@ -44,8 +45,8 @@ describe('PG Integration', () => {
 
     const umzug = new Umzug({
       migrations: umzeption({
-        dependencies: ['./fixtures/test-dependency'],
-        glob: ['fixtures/migrations/*.js'],
+        dependencies: ['../../umzeption/test/fixtures/test-dependency'],
+        glob: ['../../umzeption/test/fixtures/migrations/*.js'],
         meta: import.meta,
       }),
       context,
