+++ packages/umzeption/index.js
@@ -1,9 +1,3 @@
-// TODO(Sprint 27): Extract pg context into a separate "umzeption-pg" package (see SPRINT-PLAN.md Phase 4)
-export {
-  createUmzeptionPgContext,
-  UmzeptionPgStorage,
-} from './lib/context-pg/main.js';
-
 export {
   createUmzeptionContext,
 } from './lib/context.js';
@@ -14,7 +8,7 @@ export {
 
 export {
   installSchemaFromString,
-  pgInstallSchemaFromString,
+  registerSchemaInstaller,
 } from './lib/schema-helper.js';
 
 export {
